Unify NativeControllers
