+++ app/src/main/java/com/vlad805/fmradio/service/fm/IFMEventPoller.java
@@ -1,13 +1,15 @@
 package com.vlad805.fmradio.service.fm;
 
-import com.vlad805.fmradio.service.fm.implementation.AbstractFMController;
-
 /**
  * The interface for the controller, which constantly polls FM for updates,
  * for example, for PS, RT, etc..
  * vlad805 (c) 2020
  */
 public interface IFMEventPoller {
+	interface Callback<T> {
+		void onResult(T result);
+	}
+
 	/**
 	 * Call a poll for fresh data
 	 * @param callback Call back with Bundle fresh data
@@ -16,5 +18,5 @@ public interface IFMEventPoller {
 	 *  - C.Key.PS
 	 *  - C.Key.RT
 	 */
-	void poll(final AbstractFMController.Callback<android.os.Bundle> callback);
+	void poll(final Callback<android.os.Bundle> callback);
 }
